bugFix: app had multiple state change listeners
When copying most of this from the react-native docs, I took some liberties with changing the code. I think that introduced a state where the state listener could become added to the app more than once.

I also ran into an issue where debugging to Android, pressing home, it doesnt behave correctly, but ive seen that before -- more a function of the debugger than an issue with the app